# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de proyecto TT",
        "description": "Se implementan todos los métodos HTTP soportados por el Backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/createTipoEjercicio": {
            "post": {
                "tags": [
                    "Ejercicios"
                ],
                "summary": "Creación de un tipo de ejercicio",
                "description": "Se crea un tipo de ejercicio",
                "operationId": "0e9c1b7baed24530ef582350c2d4e6fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre_tipo": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Se almacena un tipo de ejercicio"
                    },
                    "400": {
                        "description": "Duplicidad de valores."
                    },
                    "500": {
                        "description": "Error en la base de datos"
                    }
                }
            }
        },
        "/api/getAllTipoEjercicio": {
            "get": {
                "tags": [
                    "Ejercicios"
                ],
                "summary": "Consulta los tipos de ejercicio",
                "description": "Consulta todos los tipos de ejercicios",
                "operationId": "5bff5a72cd9d0e130ffdf4cc614175bf",
                "responses": {
                    "200": {
                        "description": "Retorna una lista con los tipos de ejercicios"
                    },
                    "500": {
                        "description": "Error en la base de datos"
                    }
                }
            }
        },
        "/api/updateTypeTrining/{id_tipoEjercicio}": {
            "put": {
                "tags": [
                    "Ejercicios"
                ],
                "summary": "Actualiza una inscripción",
                "description": "Se actualiza un tipo de ejercicio",
                "operationId": "6a5e16506ccd937167e8b4036f324ca1",
                "parameters": [
                    {
                        "name": "id_tipoEjercicio",
                        "in": "path",
                        "description": "Id del tipo de ejercicio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre_tipo": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Se almacena un tipo de ejercicio"
                    },
                    "400": {
                        "description": "Duplicidad de valores."
                    },
                    "500": {
                        "description": "Error en la base de datos"
                    }
                }
            }
        },
        "/api/deleteTypeTrining/{id_tipoEjercicio}": {
            "delete": {
                "tags": [
                    "Ejercicios"
                ],
                "summary": "Elimina un tipo de ejercicio",
                "description": "Elimina un tipo de ejercicio",
                "operationId": "8a8d3424d141027245bcc32aae2f78ae",
                "parameters": [
                    {
                        "name": "id_tipoEjercicio",
                        "in": "path",
                        "description": "Id del tipo de ejercicio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Registro no encontrado"
                    },
                    "500": {
                        "description": "Error el servidor"
                    }
                }
            }
        },
        "/api/createEjercicio": {
            "post": {
                "tags": [
                    "Ejercicios"
                ],
                "summary": "Creación de un ejercicio",
                "description": "Se crea un ejercicio\n\nLa demo del ejercicio debe ser un vídeo de Youtube, se valida antes de hacer el registro",
                "operationId": "b717769d14f995363612eb1a3d92806d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_tipo_ejercicio": {
                                        "type": "integer"
                                    },
                                    "nombre_ejercicio": {
                                        "type": "string"
                                    },
                                    "unidad_medida": {
                                        "type": "string"
                                    },
                                    "demo_ejercicio": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Se almacena un tipo de ejercicio"
                    },
                    "404": {
                        "description": "Error de validación."
                    },
                    "500": {
                        "description": "Error en el servidor"
                    }
                }
            }
        },
        "/api/getAllEjercicios": {
            "get": {
                "tags": [
                    "Ejercicios"
                ],
                "summary": "Consulta los ejercicios",
                "description": "Se consultan todos los ejercicios",
                "operationId": "2cd73f1f1559c968ce6c0031e1b04e19",
                "responses": {
                    "200": {
                        "description": "Retorna una lista con los ejercicios"
                    },
                    "500": {
                        "description": "Error en la base de datos"
                    }
                }
            }
        },
        "/api/getInfoBasicEjercicios": {
            "get": {
                "tags": [
                    "Ejercicios"
                ],
                "summary": "Consulta los ejercicios",
                "description": "Se consulta la información básica de los ejercicios",
                "operationId": "26306e30d957c2009a113c21ec026905",
                "responses": {
                    "200": {
                        "description": "Retorna una lista con la información básica de los ejercicios"
                    },
                    "500": {
                        "description": "Error general"
                    }
                }
            }
        },
        "/api/updateEjercicio/{id_ejercicio}": {
            "put": {
                "tags": [
                    "Ejercicios"
                ],
                "summary": "Actualiza un ejercicio",
                "description": "Se actualiza la información de un ejercicio",
                "operationId": "696aa3d85a318fac55a8ebe6a7d7fc50",
                "parameters": [
                    {
                        "name": "id_ejercicio",
                        "in": "path",
                        "description": "Id del ejercicio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_tipo_ejercicio": {
                                        "type": "integer"
                                    },
                                    "nombre_ejercicio": {
                                        "type": "string"
                                    },
                                    "unidad_medida": {
                                        "type": "string"
                                    },
                                    "demo_ejercicio": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Se almacena un role."
                    },
                    "400": {
                        "description": "Duplicidad de valores."
                    },
                    "500": {
                        "description": "Error en la base de datos"
                    }
                }
            }
        },
        "/api/deleteEjercicio/{id_ejercicio}": {
            "delete": {
                "tags": [
                    "Ejercicios"
                ],
                "summary": "Elimina un ejercicio",
                "description": "Elimina un ejercicio",
                "operationId": "d6ef721bfc9caa0e85eb60a6148fd69a",
                "parameters": [
                    {
                        "name": "id_ejercicio",
                        "in": "path",
                        "description": "Id del ejercicio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Registro no encontrado"
                    },
                    "500": {
                        "description": "Error el servidor"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}